refactor: split app shell coordinator
